# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Lançamento e Prestação de Contas",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/categorias": {
      "get": {
        "tags": [
          "categorias"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_categorias_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incluir_inativos",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Inclui registros desativados",
              "default": false,
              "title": "Incluir Inativos"
            },
            "description": "Inclui registros desativados"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoriaOut"
                  },
                  "title": "Response Listar Api V1 Categorias Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "categorias"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_categorias_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriaIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categorias/{obj_id}": {
      "patch": {
        "tags": [
          "categorias"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_categorias__obj_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "obj_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Obj Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projetos": {
      "get": {
        "tags": [
          "projetos"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_projetos_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incluir_inativos",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Inclui registros desativados",
              "default": false,
              "title": "Incluir Inativos"
            },
            "description": "Inclui registros desativados"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjetoOut"
                  },
                  "title": "Response Listar Api V1 Projetos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projetos"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_projetos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjetoIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjetoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projetos/{obj_id}": {
      "patch": {
        "tags": [
          "projetos"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_projetos__obj_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "obj_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Obj Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjetoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjetoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/centros-custo": {
      "get": {
        "tags": [
          "centros de custo"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_centros_custo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incluir_inativos",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Inclui registros desativados",
              "default": false,
              "title": "Incluir Inativos"
            },
            "description": "Inclui registros desativados"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CentroCustoOut"
                  },
                  "title": "Response Listar Api V1 Centros Custo Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "centros de custo"
        ],
        "summary": "Criar",
        "operationId": "criar_api_v1_centros_custo_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CentroCustoIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CentroCustoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/centros-custo/{obj_id}": {
      "patch": {
        "tags": [
          "centros de custo"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_v1_centros_custo__obj_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "obj_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Obj Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CentroCustoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CentroCustoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CategoriaIn": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Nome"
          }
        },
        "type": "object",
        "required": [
          "nome"
        ],
        "title": "CategoriaIn"
      },
      "CategoriaOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "ativo"
        ],
        "title": "CategoriaOut"
      },
      "CategoriaUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "ativo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ativo"
          }
        },
        "type": "object",
        "title": "CategoriaUpdate"
      },
      "CentroCustoIn": {
        "properties": {
          "codigo": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Codigo"
          },
          "nome": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Nome"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "nome"
        ],
        "title": "CentroCustoIn"
      },
      "CentroCustoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "nome",
          "ativo"
        ],
        "title": "CentroCustoOut"
      },
...
      "CentroCustoUpdate": {
        "properties": {
          "codigo": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo"
          },
          "nome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "ativo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ativo"
          }
        },
        "type": "object",
        "title": "CentroCustoUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjetoIn": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Nome"
          }
        },
        "type": "object",
        "required": [
          "nome"
        ],
        "title": "ProjetoIn"
      },
      "ProjetoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "ativo"
        ],
        "title": "ProjetoOut"
      },
      "ProjetoUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "ativo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ativo"
          }
        },
        "type": "object",
        "title": "ProjetoUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}